fix: allocate clustering metric buffers in the input dtype
`davies_bouldin_score` and `calinski_harabasz_score` allocated their
accumulators without a `dtype`, so the buffers were float32 whatever the
input was. `dunn_index`, same family and same signature, preserves the
input dtype, so the clustering metrics disagreed with each other.

For Davies-Bouldin this was not only a dtype-label problem. Writing a
float64 centroid into a float32 buffer rounds it to roughly seven
significant digits, so `cluster_k - centroids[k]` is dominated by that
rounding once the data sits away from the origin. The score is then
wrong, not merely imprecise:

    offset   returned    correct     relative error
    0        27.011803   27.011805   1.0e-07
    1e3      27.014309   27.011805   9.3e-05
    1e5      29.370480   27.011805   8.7e-02
    1e7       1.726807   27.011806   9.4e-01

Off-origin coordinates are ordinary -- geographic positions, epoch
timestamps, unnormalised sensor readings -- and a user who reaches for
float64 because of a large dynamic range had it silently discarded.

Calinski-Harabasz narrows only the running accumulation, so its error was
bounded near float32 epsilon, but it was still a silent float32 ceiling
on a float64 input.

Both degenerate branches also returned an explicit float32 scalar; those
now follow the input dtype too.

Values are unchanged for float32 inputs, which already allocated float32
buffers.

Fixes #3467

+++ src/torchmetrics/functional/clustering/davies_bouldin_score.py
@@ -47,8 +47,10 @@ def davies_bouldin_score(data: Tensor, labels: Tensor) -> Tensor:
     num_samples, dim = data.shape
     _validate_intrinsic_labels_to_samples(num_labels, num_samples)
 
-    intra_dists = torch.zeros(num_labels, device=data.device)
-    centroids = torch.zeros((num_labels, dim), device=data.device)
+    # allocated in the input dtype: a float32 centroid buffer rounds the stored centre to ~7 significant digits, so
+    # `cluster_k - centroids[k]` catastrophically cancels for data that is not centred near the origin
+    intra_dists = torch.zeros(num_labels, device=data.device, dtype=data.dtype)
+    centroids = torch.zeros((num_labels, dim), device=data.device, dtype=data.dtype)
     for k in range(num_labels):
         cluster_k = data[labels == k, :]
         centroids[k] = cluster_k.mean(dim=0)
@@ -58,7 +60,7 @@ def davies_bouldin_score(data: Tensor, labels: Tensor) -> Tensor:
     cond1 = torch.allclose(intra_dists, torch.zeros_like(intra_dists))
     cond2 = torch.allclose(centroid_distances, torch.zeros_like(centroid_distances))
     if cond1 or cond2:
-        return torch.tensor(0.0, device=data.device, dtype=torch.float32)
+        return torch.tensor(0.0, device=data.device, dtype=data.dtype)
 
     centroid_distances[centroid_distances == 0] = float("inf")
     combined_intra_dists = intra_dists.unsqueeze(0) + intra_dists.unsqueeze(1)
